OS_OpenConsole -> Platform::WinEnsureConsole
