+++ src/Application/main.cpp
@@ -18,6 +18,10 @@ int MM_Main(int argc, char **argv) {
         Log *log = EngineIoc::ResolveLogger();
 
         std::unique_ptr<Platform> platform = Platform::CreateStandardPlatform(log);
+        platform->WinEnsureConsole();
+
+        logger->Info("World of Might and Magic, compiled: %s %s", __DATE__, __TIME__);
+        logger->Info("Extra build information: %s/%s/%s %s", BUILD_PLATFORM, BUILD_TYPE, BUILD_COMPILER, PROJECT_VERSION);
 
         std::string cmd;
         for (int i = 1; i < argc; ++i)
