+++ src/Platform/Posix/Posix.cpp
@@ -14,13 +14,14 @@
 #include <filesystem>
 
 #include "Utility/String.h"
-
-bool OS_OpenConsole() {
-    return true;
-}
+#include "Platform/Sdl/SdlPlatform.h"
 
 //////////////////// There is no Windows Registry ////////////////////
 
 bool OS_GetAppString(const char* path, char* out_string, int out_string_size) {
     return false;
+}
+
+std::unique_ptr<Platform> Platform::CreateStandardPlatform(Log *log) {
+    return std::make_unique<SdlPlatform>(log);
 }
