+++ src/Platform/Win/Win.cpp
@@ -11,14 +11,3 @@
 
 #pragma comment(lib, "winmm.lib")
 
-bool OS_OpenConsole() {
-    if (AllocConsole()) {
-        freopen("conin$", "r", stdin);
-        freopen("conout$", "w", stdout);
-        freopen("conout$", "w", stderr);
-
-        return true;
-    }
-
-    return false;
-}
