+++ test/Testing/Engine/TestPlatform.h
@@ -21,6 +21,7 @@ class TestPlatform: public Platform {
     virtual std::vector<Recti> DisplayGeometries() const override;
     virtual void ShowMessageBox(const std::string &message, const std::string& title) const override;
     virtual int64_t TickCount() const override;
+    virtual void WinEnsureConsole() const override;
 
  private:
     std::unique_ptr<Platform> base_;
